Port the transitions-objc stack to the new streams runtime.
Summary:
This is not the complete stack. Included in this diff:

- TransitionDirector type.
- Transition type.
- TransitionController and viewController.transitionDirector property.
- TransitionWindow type.

Not included in this diff:

- Gesture support.
- Transition dismissers.
- Context view retrieval.

In an effort to keep this diff to the MVP of transitions I'll be sending out the missing features as follow-up diffs. This diff represents the minimum viable code to allow someone to create a view controller transition.

Reviewers: chuga, O4 Material Apple platform reviewers, O2 Material Motion, #material_motion

Reviewed By: chuga, O4 Material Apple platform reviewers

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2394