Add support for Core Animation in MotionObservable.
Summary:
This diff introduces a new channel to MotionObservable for emitting Core Animation objects. Only certain operators and properties support core animation, with all others throwing assertions at runtime.

This diff solves the following problems:

- Being able to use core animation as a source for tweens and springs interchangeably with T-value emitters (such as pop).
- Being able to apply map transformations to core animation streams. E.g. it's possible to extract the .x() from a CGPoint animation.

That core animation support is a runtime consideration and not a compile time one is still an open problem to resolve, but this diff is a strong step forward toward providing first-class core animation support in the ecosystem.

Reviewers: chuga, O4 Material Apple platform reviewers, O2 Material Motion, #material_motion

Reviewed By: chuga, O4 Material Apple platform reviewers

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2416