Add support for removing in-flight Core Animation animations.
Summary:
The Core Animation channel now accepts a Swift enum type of either .add or .remove.

When an animation is removed, we extract the latest presentation layer value and write it to the model layer. This may bite us in the future, but at least for the examples we currently have this is nice expected behavior.

Reviewers: chuga, O4 Material Apple platform reviewers, O2 Material Motion, #material_motion

Reviewed By: chuga, O4 Material Apple platform reviewers

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2423