Emit the last known state when a property is subscribed to.
Summary:
There is an unwritten rule that sources are expected to emit their "initial state" when subscribed to. Properties can technically be sources, so we need to ensure that they adhere to this guarantee.

This reveals a rather unfortunate fact: writing multiple streams to a single property will yield undefined "last-writer wins" behaviors. In general we shouldn't be writing multiple streams to the same property, so this may be something we want to throw an assertion on at runtime and encourage the use of combinator operators such as toggled.

Reviewers: chuga, O4 Material Apple platform reviewers, O2 Material Motion, #material_motion

Reviewed By: chuga, O4 Material Apple platform reviewers

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2444