Spring now owns its own MotionObservable.
Summary:
This diff was initiated as an exploration into removing the Plan and Source type from the ecosystem. An earlier version of this diff turned Spring into a SpringObservable (a subclass of MotionObservable).  Didn't pursue this because introducing new subtypes of MotionObservable doesn't feel scalable and is generally against our "avoid subclassing" philosophy.

Instead, the Spring type now creates and owns its own valueStream instance. This creates a clear relationship between a Spring and its stream of values. This will position the Spring type to eventually clearly communicate that subscribing multiple properties to a single spring will result in the same values being broadcast to all. To achieve this goal we'll need to `memoize` the valueStream before we store it on the spring. This will be achieved in a follow-up diff after the `memoize` operator is introduced.

Reviewers: O4 Material Apple platform reviewers, O2 Material Motion, #material_motion, appsforartists

Reviewed By: O2 Material Motion, #material_motion, appsforartists

Subscribers: appsforartists

Tags: #material_motion

Differential Revision: http://codereview.cc/D2482