Reactive properties now create and own their own stream.
Summary:
Properties are now a distinct type rather than a "pseudo-stream" type (properties used to expose a subscribe method). In a follow-up change I'll introduce MotionObservableConvertible which will make it possible to use properties in the same places that streams are used, removing the need for `property.stream`.

Depends on D2560

Reviewers: chuga, O4 Material Apple platform reviewers, O2 Material Motion, #material_motion

Reviewed By: chuga, O4 Material Apple platform reviewers

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2561