Add a sticker picker example.
Summary:
This is a first pass at the example. A future iteration will make it so that a sticker can be added by long-pressing the sticker in the picker. The long-press will dismiss the picker view controller and seamlessly move to a directly manipulable interaction which will end with the sticker being placed.

This example could also benefit from demoing how to react to a sticker's location changing in order to demonstrate how to connect interactions to business logic.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2643