Spring shape updated to better support enabling/disabling and state observation.
Summary:
Spring now has both a state and enabled property. The Spring instance's value stream is multicast, and the underlying system that drives this value stream is expected to update the spring's state property.

This is a precursor to being able to delete the state channel in its entirety. I'll be cleaning up the mechanism by which the spring's value stream is created in a follow-up change.

In order to keep the examples functioning, this change also introduces "terminal conditions" for transitions. This is a hack solution and will need to be iterated on.

Reviewers: O4 Material Apple platform reviewers, appsforartists, #material_motion, O2 Material Motion

Reviewed By: appsforartists, #material_motion, O2 Material Motion

Subscribers: appsforartists

Tags: #material_motion

Differential Revision: http://codereview.cc/D2683