Expose threshold on TransitionSpring and provide threshold throughout all examples.
Summary:
We were defaulting the threshold to 1 which was breaking non-positional springs when used with POP. All examples have been updated to use correct threshold values.

Ideally the L1 user won't have to worry about threshold so this API could use a revisiting. For now this at least brings the examples up to a reasonable level of correctness.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, appsforartists

Reviewed By: O2 Material Motion, #material_motion, appsforartists

Tags: #material_motion

Differential Revision: http://codereview.cc/D2690