Add a completion block to the core animation channel.
Summary:
This ensures that the completion block is associated with the animation that's ultimately committed to the property.

We were previously depending on the assumption that the animation would be added to a layer when we emitted an animation down the core animation channel. This isn't always the case when operators are in play, e.g. multicast, which caches the last emitted value and then emits it to any subscribed observers. This change ensures that the completion handlers are always registered.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Subscribers: appsforartists

Tags: #material_motion

Differential Revision: http://codereview.cc/D2692