Fix memory leak in Spring.
Summary:
Spring was providing itself to the system, which in turn was returning an observable that held a reference of the spring, which in turn was stored on the spring. This created a retain cycle.

This diff breaks the retain cycle by providing the system with a shadow copy of the spring instance. This makes it easy for the system implementations to do the "right thing" and resolves the retain cycle.

This solution also makes it possible to expose spring.state as a stream rather than as a reactive property. This ensures that only the spring system is able to write to the spring's state.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2704