Clean up the MotionObservable implementation and improve test coverage.
Summary:
I've made the CoreAnimation channel optional in MotionObserver as an experiment to see if this leads to a better solution for identifying when Core Animation is supported vs when it's not.

Also making use of the `subscribe(observer: observer)` short-hand where applicable throughout the code base.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2740