Simplify the modal dialog example using Tossable and TransitionSpring.
Summary:
The previous implementation was essentially rewriting Tossable. This change refactors the interaction to make use of the Tossable interaction by providing a TransitionSpring to Tossable. This takes advantage of the fact that TransitionSpring is a subclass of Spring and that Tossable can be provided with an existing Spring instance.

Note that this diff makes it possible to drag the modal dialog on the x-axis. In a follow-up diff I will introduce constraints so that this interaction can be locked to the y axis.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2817