Introduce runtime.connect.
Summary:
This API is used to connect a single stream to a reactive property. `runtime.add` is now only used for adding interactions to properties or views.

This separation of registration mechanisms will help us audit code for use of streams throughout L1-space. It also paints a clear picture of when an interaction is being used vs a stream or property.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2824