Add constraint support to additive animations.
Summary:
Additive animation values are now calculated just before the animation is committed to the layer. This allows constraints to treat the animation from/to values as absolute values, rather than relative ones.

This change also introduces some light refactoring of the Core Animation add information. The information is now stored in a documented structure instead of an inline tuple, making it easier to copy, modify, and initialize the information.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2833