Move all examples to the Catalog target and resolve all API visibility failures.
Summary:
Generating the examples using CocoaPods caused the examples to be placed in the same build unit as the library source, giving the examples access to all `internal` APIs.

The examples are now in the app target. This change revealed a number of APIs that should have been public that were marked internal.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2840