Adjust MotionObservable subscription APIs to encourage forwarding channels rather than creating a terminal subscription.
Summary:
The short-form subscribe(next:) method causes a stream to terminate its channels. This can be problematic for operators that are expected to pass along their core animation and visualization channels.

This diff adjusts the APIs to encourage passing the observer along with subscribe when relevant, and to use `_subscribe` when a terminal subscription is intentionally being created.

We want to discourage use of _subscribe in app space, so I've prefixed the API with a _ to indicate that fact.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2866