Fix bug with runtime.get and gesture recognizer typing.
Summary:
The type-agnostic object cache crashes when working with generic types such as ReactiveUIGestureRecognizer because we're attempting to do a safe cast to the desired type. This breaks when attempting to cast from a ReactiveUIGestureRecognize<UIGestureRecognizer> to a ReactiveUIGestureRecognize<UIPanGestureRecognizer>, for example.

We now use an unsafe bit cast to force the conversion.

This diff fixes the directly manipulable example, which was crashing before this diff.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Subscribers: chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2875