Trim down the arc move example's code and move most logic to the example source.
Summary:
TimelineView now uses a delegate + observation pattern so that the timeline logic can stay in the example source.

Also removed the dynamic time connection for now in order to keep the example lean. We can introduce a separate example in the future that shows how to do dynamic tween animation timing.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2893