Fix bug causing Tween never to come to rest.
Summary:
Tween is now using a checkAndEmit model to wait until it has received all of its necessary configuration information before emitting an animation. If all of this information is available at subscription time then we'll be able to emit the animation right away. If we subscribe to enabled after all of the properties, this means that we could emit the animation twice in rapid succession. This appears to break the Core Animation completion handler and we're left with animations that never come to rest.

The resolution here is to subscribe to the enabled state before subscribing to all of the configuration values.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2918