Reintroduce property core animation event forwarding.
Summary:
This was previously removed because it caused weird behavior when connecting unrelated property types, e.g. position to scale. When connecting properties of similar types however, this behavior is beneficial.

The solve for the unrelated properties case is to make use of valve to keep animations from propagating to the other property, as seen in the interactive push back transition example.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2956