Fix bug causing properties-chained-to-properties to not animate correctly.
Summary:
The Core Animation systems choose a unique key to refer to the animation they're emitting so that if they need to remove the animation in the future they can do so. This key is used when adding the animation to the layer.

When connecting properties together, there is only one system involved and therefor one key. This causes animations to be added with the same key, which causes the previously-added animations on the same key to be replaced.

This patch fixes this bug by uniquing the animation keys with the keyPath.

Reviewers: O4 Material Apple platform reviewers, O2 Material Motion, #material_motion, markwei

Reviewed By: O4 Material Apple platform reviewers, O2 Material Motion, #material_motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D2976