Add foreAlignmentEdge property to TransitionController.
Summary:
If a view controller has a non-zero preferredContentSize, then the foreAlignmentEdge will be used to align the view to either the center of the screen, if nil, or to the specified edge (one of minX, minY, maxX, or maxY).

This API is useful for building modal dialogs and sliding drawers that are presented over the current context with `modalPresentationStyle = .overCurrentContext`.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion

Tags: #material_motion

Differential Revision: http://codereview.cc/D3067