When no gesture recognizer is provided to a gestural interaction that expects one, the interaction now does nothing.
Summary:
Gestural interactions configured to expect a gesture recognizer will no longer connect any streams if a gesture recognizer is not provided.

The `withFirstGestureIn:` now uses this new behavior when it can't find a gesture recognizer that qualifies for the interaction. Prior to this change, if the withFirstGestureIn initializer couldn't find a gesture recognizer it would default to `registerNewRecognizerToTargetView` behavior. This can cause superfluous gesture recognizers to be registered to a view.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3077