Add a Manipulation type and implement UIKit view controller transitioning interactivity APIs.
Summary:
Draggable, Rotatable, and Scalable are now Manipulation types. When added to the runtime they will now affect the runtime's isBeingManipulated stream.

We use this stream to inform UIKit when a transition becomes or stops being interactive. This ensures that system animations (e.g. status bar) start when the user stops interacting with the transition.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3094