Add a resistance property to Draggable.
Summary:
This new property makes it possible to configure resistance behaviors on Draggable when the user moves beyond a given perimeter.

This API is preferred over applying a rubberBand constraint to a Tossable interaction because it ensures that resistance is only applied to the draggable output, not the spring (which may re-read the position value and then incorrectly cause the position to jump to the doubly-resisted position).

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3097