Add a defaultModalPresentationStyle API for transitions with presentation.
Summary: This allows transitions with presentation to choose the default modal presentation style for a view controller. This is necessary because the transition may choose not to use a presentation controller even though it conforms to the TransitionWithPresentation protocol (e.g. a modal transition that covers the whole screen may not need a presentation controller).

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Reviewed By: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, randcode-generator, markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D3230